# specs/1-ai-native-book/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-Native Book API
  description: API for the AI-Native Book on Physical AI & Humanoid Robotics
  version: 1.0.0
servers:
  - url: https://api.example.com
    description: Production server
  - url: https://staging-api.example.com
    description: Staging server
paths:
...
  /api/search:
    post:
      summary: Search book content
      description: Search across all book content using the query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '500':
          description: Internal server error
  /api/chat:
    post:
      summary: Chat with the RAG system
      description: Process user queries against book content using RAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '500':
          description: Internal server error
...
components:
  schemas:
...
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Search query string
        filters:
          type: object
          properties:
            module:
              type: string
              description: Filter by specific module
            chapter:
              type: string
              description: Filter by specific chapter
            tags:
              type: array
              items:
                type: string
              description: Filter by tags
          description: Optional filters for the search
      required:
        - query
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Identifier of the result
              title:
                type: string
                description: Title of the result
              contentPreview:
                type: string
                description: Preview of the content
              module:
                type: string
                description: Module containing the result
              chapter:
                type: string
                description: Chapter containing the result
              relevanceScore:
                type: number
                description: Relevance score of the result
        totalCount:
          type: integer
          description: Total number of results
      required:
        - results
        - totalCount
    ChatRequest:
      type: object
      properties:
        message:
          type: string
          description: User's message/query
        context:
          type: object
          properties:
            moduleId:
              type: string
              description: Module context for the query
            chapterId:
              type: string
              description: Chapter context for the query
          description: Context information for the query
        sessionId:
          type: string
          description: Chat session identifier
      required:
        - message
    ChatResponse:
      type: object
      properties:
        response:
          type: string
          description: AI-generated response
        sources:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Source identifier
              title:
                type: string
                description: Source title
              module:
                type: string
                description: Module containing the source
              chapter:
                type: string
                description: Chapter containing the source
              relevance:
                type: number
                description: Relevance score of the source
          description: Sources used in the response
        sessionId:
          type: string
          description: Chat session identifier
      required:
        - response
        - sources
        - sessionId
...